Moved input validation tests into their own class.
